update CBC code
Updated the CBC code not to manipulate the splices directly.
I found this way to implement it clearer in showing how CBC
works.
It's probably slower, but, right now performance is not an issue.
Less hard-core too, I suppose, but that's also not an issue; clear
over clever and all that...