add solution set 2 challenge 12
Possibly I should add a better solution for padding.
The padding is a problem for the lookup table because it changes
the plaintext.
So we have a known plaintext ending in \x01 and after that
the oracle is processing a plaintext ending in \x02\x02... so
the ciphertexts won't match.

And what if the Oracle is padding in a weird way?