update pkcs#7 code
As suggested in challenge 15 I have updated the PKCS#7 code.
That also revelaed a few flaws in the implementations and tests,
that I, hopefully, fixed.

For reason that I have not investigated that also revealed a big
logical error in the ch13 code (was assuming the role to be 'use'
instead or 'user').