update PKCS and AES code
There was an issue with blindly applying PKCS#7 rule to the
plaintext.
AES cipher doesn't care how you pad the plaintext, because it
only cares about 16bytes blocks, and how you go from the original
plaintext to one multiple of 16 bytes long and vice versa doesn't
really concern it.

PKCS#7 (smartly) requires that if the plaintext is of the right
size you add an extra block. This makes easy and to remove the
padding and removes ambiguous cases (e.g. last byte is \x01).

But, for my scope, this just makes things more complicated;
because I may end up with an additional ciphertext block when I
don't really expect it.
It's not something I cannot work around, but it just makes the
code more complex.

So, I tried to not implement the PKC#7 correctly when I encrypt,
but that broke other stuff and it to fix them I had to write
more complex in any case.

Not sure this is the right way, matter of fact.

A part of me thinks I should follow the PKCS#7 rules, but, again,
AES doesn't say anything about padding, and, for what I know, I
could be using any kind of padding, even a broken one as I am doing.

I guess, I'll keep going and see how it works out.