update challenge #17 solution
If the last block was padded, it would mess up with the attack.
So, I wrote a routine to exploit the fact that if we know that
some bytes of the block are padded, we can easily guess the
plaintext of those bytes (is the padding value) and calculate
D2 for those bytes.
The rest of the block is processed as before.